requirements.txt
pytesseract==0.3.8
Pillow==8.3.2